RF: move specific search mode implementations into individual files
It became virtually impossible to orient myself in a single file with multipl
implementations.  This RF is long overdue, thus  done here.  No functionality
changes should be in this commit - only code moves
