+++ datalad/metadata/search.py
@@ -11,1074 +11,24 @@
 
 __docformat__ = 'restructuredtext'
 
-import collections
 import logging
-from datalad.log import log_progress
 lgr = logging.getLogger('datalad.metadata.search')
 
-import os
-import re
-from functools import partial
-from os.path import join as opj, exists
-from os.path import relpath
-from os.path import normpath
-import sys
-from time import time
+from datalad.metadata.search_base import _search_from_virgin_install
+from datalad.metadata.search_egrep import _EGrepCSSearch, _EGrepSearch
+from datalad.metadata.search_pyeval import _PyEvalSearch
+from datalad.metadata.search_whoosh import _BlobSearch, _AutofieldSearch
 
-from datalad import cfg
 from datalad.interface.base import Interface
 from datalad.interface.base import build_doc
 from datalad.interface.utils import eval_results
-from datalad.distribution.dataset import Dataset
 from datalad.distribution.dataset import datasetmethod, EnsureDataset, \
     require_dataset
-from datalad.support.gitrepo import GitRepo
 from datalad.support.param import Parameter
 from datalad.support.constraints import EnsureNone
 from datalad.support.constraints import EnsureInt
 
-from datalad.consts import SEARCH_INDEX_DOTGITDIR
-from datalad.utils import (
-    as_unicode,
-    assure_list,
-    assure_unicode,
-    get_suggestions_msg,
-    shortened_repr,
-    unicode_srctypes,
-)
 from datalad.support.exceptions import NoDatasetFound
-from datalad.ui import ui
-from datalad.dochelpers import single_or_plural
-from datalad.dochelpers import exc_str
-from datalad.metadata.metadata import query_aggregated_metadata
-
-# TODO: consider using plain as_unicode, without restricting
-# the types?
-_any2unicode = partial(as_unicode, cast_types=(int, float, tuple, list, dict))
-
-
-def _listdict2dictlist(lst, strict=True):
-    """Helper to deal with DataLad's unique value reports
-
-    Parameters
-    ----------
-    lst : list
-      List of dicts
-    strict : bool
-      In strict mode any dictionary items that doesn't have a simple value
-      (but another container) is discarded. In non-strict mode, arbitrary
-      levels of nesting are preserved, and no unique-ification of values
-      is performed. The latter can be used when it is mostly (or merely)
-      interesting to get a list of metadata keys.
-    """
-    # unique values that we got, always a list
-    if all(not isinstance(uval, dict) for uval in lst):
-        # no nested structures, take as is
-        return lst
-
-    type_excluder = (tuple, list)
-    if strict:
-        type_excluder += (dict,)
-    # we need to turn them inside out, instead of a list of
-    # dicts, we want a dict where values are lists, because we
-    # cannot handle hierarchies in a tabular search index
-    # if you came here to find that out, go ahead and use a graph
-    # DB/search
-    udict = {}
-    for uv in lst:
-        if not isinstance(uv, dict):
-            # we cannot mix and match, discard any scalar values
-            # in favor of the structured metadata
-            continue
-        for k, v in uv.items():
-            if isinstance(v, type_excluder):
-                # this is where we draw the line, two levels of
-                # nesting. whoosh can only handle string values
-                # injecting a stringified blob of something doesn't
-                # really enable anything useful -> graph search
-                continue
-            if v == "":
-                # no cruft
-                continue
-            if strict:
-                # no duplicate values, only hashable stuff
-                uvals = udict.get(k, set())
-                uvals.add(v)
-            else:
-                # whatever it is, we'll take it
-                uvals = udict.get(k, [])
-                uvals.append(v)
-            udict[k] = uvals
-    return {
-        # set not good for JSON, have plain list
-        k: list(v) if len(v) > 1 else list(v)[0]
-        for k, v in udict.items()
-        # do not accumulate cruft
-        if len(v)}
-
-
-def _meta2autofield_dict(meta, val2str=True, schema=None, consider_ucn=True):
-    """Takes care of dtype conversion into unicode, potential key mappings
-    and concatenation of sequence-type fields into CSV strings
-
-    - if `consider_ucn` (default) it would copy keys from
-      datalad_unique_content_properties into `meta` for that extractor
-    - ... TODO ...
-    """
-    if consider_ucn:
-        # loop over all metadata sources and the report of their unique values
-        ucnprops = meta.get("datalad_unique_content_properties", {})
-        for src, umeta in ucnprops.items():
-            srcmeta = meta.get(src, {})
-            for uk in umeta:
-                if uk in srcmeta:
-                    # we have a real entry for this key in the dataset metadata
-                    # ignore any generated unique value list in favor of the
-                    # tailored data
-                    continue
-                srcmeta[uk] = _listdict2dictlist(umeta[uk], strict=False) if umeta[uk] is not None else None
-            if src not in meta and srcmeta:
-                meta[src] = srcmeta  # assign the new one back
-
-    srcmeta = None   # for paranoids to avoid some kind of manipulation of the last
-
-    def _deep_kv(basekey, dct):
-        """Return key/value pairs of any depth following a rule for key
-        composition
-
-        dct must be a dict
-        """
-        for k, v in dct.items():
-            if (k != '@id' and k.startswith('@')) or k == 'datalad_unique_content_properties':
-                # ignore all JSON-LD specials, but @id
-                continue
-            # TODO `k` might need remapping, if another key was already found
-            # with the same definition
-            key = u'{}{}'.format(
-                basekey,
-                # replace now special chars, and avoid spaces
-                # `os.sep` needs to go, because whoosh uses the field name for
-                # temp files during index staging, and trips over absent "directories"
-                # TODO maybe even kill parentheses
-                # TODO actually, it might be better to have an explicit whitelist
-                k.lstrip('@').replace(os.sep, '_').replace(' ', '_').replace('-', '_').replace('.', '_').replace(':', '-')
-            )
-            if isinstance(v, list):
-                v = _listdict2dictlist(v)
-
-            if isinstance(v, dict):
-                # dive
-                for i in _deep_kv('{}.'.format(key), v):
-                    yield i
-            else:
-                yield key, v
-
-    return {
-        k:
-        # turn lists into space-separated value strings
-            (u' '.join(_any2unicode(i) for i in v) if isinstance(v, (list, tuple)) else
-            # and the rest into unicode
-            _any2unicode(v)) if val2str else v
-        for k, v in _deep_kv('', meta or {})
-        # auto-exclude any key that is not a defined field in the schema (if there is
-        # a schema
-        if schema is None or k in schema
-    }
-
-
-def _search_from_virgin_install(dataset, query):
-    #
-    # this is to be nice to newbies
-    #
-    exc_info = sys.exc_info()
-    if dataset is None:
-        if not ui.is_interactive:
-            raise NoDatasetFound(
-                "No DataLad dataset found. Specify a dataset to be "
-                "searched, or run interactively to get assistance "
-                "installing a queriable superdataset."
-            )
-        # none was provided so we could ask user whether he possibly wants
-        # to install our beautiful mega-duper-super-dataset?
-        # TODO: following logic could possibly benefit other actions.
-        DEFAULT_DATASET_PATH = cfg.obtain('datalad.locations.default-dataset')
-        if os.path.exists(DEFAULT_DATASET_PATH):
-            default_ds = Dataset(DEFAULT_DATASET_PATH)
-            if default_ds.is_installed():
-                if ui.yesno(
-                    title="No DataLad dataset found at current location",
-                    text="Would you like to search the DataLad "
-                         "superdataset at %r?"
-                          % DEFAULT_DATASET_PATH):
-                    pass
-                else:
-                    raise exc_info[1]
-            else:
-                raise NoDatasetFound(
-                    "No DataLad dataset found at current location. "
-                    "The DataLad superdataset location %r exists, "
-                    "but does not contain an dataset."
-                    % DEFAULT_DATASET_PATH)
-        elif ui.yesno(
-                title="No DataLad dataset found at current location",
-                text="Would you like to install the DataLad "
-                     "superdataset at %r?"
-                     % DEFAULT_DATASET_PATH):
-            from datalad.api import install
-            default_ds = install(DEFAULT_DATASET_PATH, source='///')
-            ui.message(
-                "From now on you can refer to this dataset using the "
-                "label '///'"
-            )
-        else:
-            raise exc_info[1]
-
-        lgr.info(
-            "Performing search using DataLad superdataset %r",
-            default_ds.path
-        )
-        for res in default_ds.search(query):
-            yield res
-        return
-    else:
-        raise  # this function is called within exception handling block
-
-
-class _Search(object):
-    def __init__(self, ds, **kwargs):
-        self.ds = ds
-        self.documenttype = self.ds.config.obtain(
-            'datalad.search.index-{}-documenttype'.format(self._mode_label),
-            default=self._default_documenttype)
-
-    def __call__(self, query, max_nresults=None):
-        raise NotImplementedError
-
-    @classmethod
-    def _key_matches(cls, k, regexes):
-        """Return which regex the key matches
-        """
-        for regex in regexes:
-            if re.search(regex, k):
-                return regex
-
-    def show_keys(self, *args, **kwargs):
-        raise NotImplementedError(args)
-
-    def get_query(self, query):
-        """Prepare query structure specific for a search backend.
-
-        It can also memorize within instance some parameters of the last query
-        which could be used to assist output formatting/structuring later on
-        """
-        raise NotImplementedError
-
-    def get_nohits_msg(self):
-        """Given what it knows, provide recommendation in the case of no hits"""
-        return "No search hits, wrong query? " \
-               "See 'datalad search --show-keys name' for known keys " \
-               "and 'datalad search --help' on how to prepare your query."
-
-
-class _WhooshSearch(_Search):
-    def __init__(self, ds, force_reindex=False, **kwargs):
-        super(_WhooshSearch, self).__init__(ds, **kwargs)
-
-        self.idx_obj = None
-        # where does the bunny have the eggs?
-
-        self.index_dir = opj(str(self.ds.repo.dot_git), SEARCH_INDEX_DOTGITDIR)
-        self._mk_search_index(force_reindex)
-
-    def show_keys(self, mode, regexes=None):
-        """
-
-        Parameters
-        ----------
-        mode: {"name"}
-        regexes: list of regex
-          Which keys to bother working on
-        """
-        if mode != 'name':
-            raise NotImplementedError(
-                "ATM %s can show only names, so please use show_keys with 'name'"
-                % self.__class__.__name__
-            )
-        for k in self.idx_obj.schema.names():
-            if regexes and not self._key_matches(k, regexes):
-                continue
-            print(u'{}'.format(k))
-
-    def get_query(self, query):
-        # parse the query string
-        self._mk_parser()
-        # for convenience we accept any number of args-words from the
-        # shell and put them together to a single string here
-        querystr = ' '.join(assure_list(query))
-        # this gives a formal whoosh query
-        wquery = self.parser.parse(querystr)
-        return wquery
-
-    def _meta2doc(self, meta, val2str=True, schema=None):
-        raise NotImplementedError
-
-    def _mk_schema(self):
-        raise NotImplementedError
-
-    def _mk_parser(self):
-        raise NotImplementedError
-
-    def _mk_search_index(self, force_reindex):
-        """Generic entrypoint to index generation
-
-        The actual work that determines the structure and content of the index
-        is done by functions that are passed in as arguments
-
-        `meta2doc` - must return dict for index document from result input
-        """
-        from whoosh import index as widx
-        from .metadata import get_ds_aggregate_db_locations
-        dbloc, db_base_path = get_ds_aggregate_db_locations(self.ds)
-        # what is the lastest state of aggregated metadata
-        metadata_state = self.ds.repo.get_last_commit_hexsha(relpath(dbloc, start=self.ds.path))
-        # use location common to all index types, they would all invalidate
-        # simultaneously
-        stamp_fname = opj(self.index_dir, 'datalad_metadata_state')
-        index_dir = opj(self.index_dir, self._mode_label)
-
-        if (not force_reindex) and \
-                exists(index_dir) and \
-                exists(stamp_fname) and \
-                open(stamp_fname).read() == metadata_state:
-            try:
-                # TODO check that the index schema is the same
-                # as the one we would have used for reindexing
-                # TODO support incremental re-indexing, whoosh can do it
-                idx = widx.open_dir(index_dir)
-                lgr.debug(
-                    'Search index contains %i documents',
-                    idx.doc_count())
-                self.idx_obj = idx
-                return
-            except widx.LockError as e:
-                raise e
-            except widx.IndexError as e:
-                # Generic index error.
-                # we try to regenerate
-                lgr.warning(
-                    "Cannot open existing index %s (%s), will regenerate",
-                    index_dir, exc_str(e)
-                )
-            except widx.IndexVersionError as e:  # (msg, version, release=None)
-                # Raised when you try to open an index using a format that the
-                # current version of Whoosh cannot read. That is, when the index
-                # you're trying to open is either not backward or forward
-                # compatible with this version of Whoosh.
-                # we try to regenerate
-                lgr.warning(exc_str(e))
-                pass
-            except widx.OutOfDateError as e:
-                # Raised when you try to commit changes to an index which is not
-                # the latest generation.
-                # this should not happen here, but if it does ... KABOOM
-                raise
-            except widx.EmptyIndexError as e:
-                # Raised when you try to work with an index that has no indexed
-                # terms.
-                # we can just continue with generating an index
-                pass
-            except ValueError as e:
-                if 'unsupported pickle protocol' in str(e):
-                    lgr.warning(
-                        "Cannot open existing index %s (%s), will regenerate",
-                        index_dir, exc_str(e)
-                    )
-                else:
-                    raise
-
-        lgr.info('{} search index'.format(
-            'Rebuilding' if exists(index_dir) else 'Building'))
-
-        if not exists(index_dir):
-            os.makedirs(index_dir)
-
-        # this is a pretty cheap call that just pull this info from a file
-        dsinfo = self.ds.metadata(
-            get_aggregates=True,
-            return_type='list',
-            result_renderer='disabled')
-
-        self._mk_schema(dsinfo)
-
-        idx_obj = widx.create_in(index_dir, self.schema)
-        idx = idx_obj.writer(
-            # cache size per process
-            limitmb=cfg.obtain('datalad.search.indexercachesize'),
-            # disable parallel indexing for now till #1927 is resolved
-            ## number of processes for indexing
-            #procs=multiprocessing.cpu_count(),
-            ## write separate index segments in each process for speed
-            ## asks for writer.commit(optimize=True)
-            #multisegment=True,
-        )
-
-        # load metadata of the base dataset and what it knows about all its subdatasets
-        # (recursively)
-        old_idx_size = 0
-        old_ds_rpath = ''
-        idx_size = 0
-        log_progress(
-            lgr.info,
-            'autofieldidxbuild',
-            'Start building search index',
-            total=len(dsinfo),
-            label='Building search index',
-            unit=' Datasets',
-        )
-        for res in query_aggregated_metadata(
-                reporton=self.documenttype,
-                ds=self.ds,
-                aps=[dict(path=self.ds.path, type='dataset')],
-                # MIH: I cannot see a case when we would not want recursion (within
-                # the metadata)
-                recursive=True):
-            # this assumes that files are reported after each dataset report,
-            # and after a subsequent dataset report no files for the previous
-            # dataset will be reported again
-            meta = res.get('metadata', {})
-            doc = self._meta2doc(meta)
-            admin = {
-                'type': res['type'],
-                'path': relpath(res['path'], start=self.ds.path),
-            }
-            if 'parentds' in res:
-                admin['parentds'] = relpath(res['parentds'], start=self.ds.path)
-            if admin['type'] == 'dataset':
-                if old_ds_rpath:
-                    lgr.debug(
-                        'Added %s on dataset %s',
-                        single_or_plural(
-                            'document',
-                            'documents',
-                            idx_size - old_idx_size,
-                            include_count=True),
-                        old_ds_rpath)
-                log_progress(lgr.info, 'autofieldidxbuild',
-                             'Indexed dataset at %s', old_ds_rpath,
-                             update=1, increment=True)
-                old_idx_size = idx_size
-                old_ds_rpath = admin['path']
-                admin['id'] = res.get('dsid', None)
-
-            doc.update({k: assure_unicode(v) for k, v in admin.items()})
-            lgr.debug("Adding document to search index: {}".format(doc))
-            # inject into index
-            idx.add_document(**doc)
-            idx_size += 1
-
-        if old_ds_rpath:
-            lgr.debug(
-                'Added %s on dataset %s',
-                single_or_plural(
-                    'document',
-                    'documents',
-                    idx_size - old_idx_size,
-                    include_count=True),
-                old_ds_rpath)
-
-        lgr.debug("Committing index")
-        idx.commit(optimize=True)
-        log_progress(
-            lgr.info, 'autofieldidxbuild', 'Done building search index')
-
-        # "timestamp" the search index to allow for automatic invalidation
-        with open(stamp_fname, 'w') as f:
-            f.write(metadata_state)
-
-        lgr.info('Search index contains %i documents', idx_size)
-        self.idx_obj = idx_obj
-
-    def __call__(self, query, max_nresults=None, force_reindex=False, full_record=False):
-        if max_nresults is None:
-            # mode default
-            max_nresults = 20
-
-        with self.idx_obj.searcher() as searcher:
-            wquery = self.get_query(query)
-
-            # perform the actual search
-            hits = searcher.search(
-                wquery,
-                terms=True,
-                limit=max_nresults if max_nresults > 0 else None)
-            # report query stats
-            topstr = '{} top {}'.format(
-                max_nresults,
-                single_or_plural('match', 'matches', max_nresults)
-            )
-            lgr.info('Query completed in {} sec.{}'.format(
-                hits.runtime,
-                ' Reporting {}.'.format(
-                    ('up to ' + topstr)
-                    if max_nresults > 0
-                    else 'all matches'
-                )
-                if not hits.is_empty()
-                else ' No matches.'
-            ))
-
-            if not hits:
-                return
-
-            nhits = 0
-            annotated_hits = []
-            # annotate hits for full metadata report
-            for i, hit in enumerate(hits):
-                annotated_hit = dict(
-                    path=normpath(opj(self.ds.path, hit['path'])),
-                    query_matched={assure_unicode(k): assure_unicode(v)
-                                   if isinstance(v, unicode_srctypes) else v
-                                   for k, v in hit.matched_terms()},
-                    parentds=normpath(
-                        opj(self.ds.path, hit['parentds'])) if 'parentds' in hit else None,
-                    type=hit.get('type', None))
-                if not full_record:
-                    annotated_hit.update(
-                        refds=self.ds.path,
-                        action='search',
-                        status='ok',
-                        logger=lgr,
-                    )
-                    yield annotated_hit
-                else:
-                    annotated_hits.append(annotated_hit)
-                nhits += 1
-            if full_record:
-                for res in query_aggregated_metadata(
-                        # type is taken from hit record
-                        reporton=None,
-                        ds=self.ds,
-                        aps=annotated_hits,
-                        # never recursive, we have direct hits already
-                        recursive=False):
-                    res.update(
-                        refds=self.ds.path,
-                        action='search',
-                        status='ok',
-                        logger=lgr,
-                    )
-                    yield res
-
-            if max_nresults and nhits == max_nresults:
-                lgr.info(
-                    "Reached the limit of {}, there could be more which "
-                    "were not reported.".format(topstr)
-                )
-
-
-class _BlobSearch(_WhooshSearch):
-    _mode_label = 'textblob'
-    _default_documenttype = 'datasets'
-
-    def _meta2doc(self, meta):
-        # coerce the entire flattened metadata dict into a comma-separated string
-        # that also includes the keys
-        return dict(meta=u', '.join(
-            u'{}: {}'.format(k, v)
-            for k, v in _meta2autofield_dict(
-                meta,
-                val2str=True,
-                schema=None).items()))
-
-    def _mk_schema(self, dsinfo):
-        from whoosh import fields as wf
-        from whoosh.analysis import StandardAnalyzer
-
-        # TODO support some customizable mapping to homogenize some metadata fields
-        # onto a given set of index keys
-        self.schema = wf.Schema(
-            id=wf.ID,
-            path=wf.ID(stored=True),
-            type=wf.ID(stored=True),
-            parentds=wf.ID(stored=True),
-            meta=wf.TEXT(
-                stored=False,
-                analyzer=StandardAnalyzer(minsize=2))
-        )
-
-    def _mk_parser(self):
-        from whoosh import qparser as qparse
-
-        # use whoosh default query parser for now
-        parser = qparse.QueryParser(
-            "meta",
-            schema=self.idx_obj.schema
-        )
-        parser.add_plugin(qparse.FuzzyTermPlugin())
-        parser.remove_plugin_class(qparse.PhrasePlugin)
-        parser.add_plugin(qparse.SequencePlugin())
-        self.parser = parser
-
-
-class _AutofieldSearch(_WhooshSearch):
-    _mode_label = 'autofield'
-    _default_documenttype = 'all'
-
-    def _meta2doc(self, meta):
-        return _meta2autofield_dict(meta, val2str=True, schema=self.schema)
-
-    def _mk_schema(self, dsinfo):
-        from whoosh import fields as wf
-        from whoosh.analysis import SimpleAnalyzer
-
-        # haven for terms that have been found to be undefined
-        # (for faster decision-making upon next encounter)
-        # this will harvest all discovered term definitions
-        definitions = {
-            '@id': 'unique identifier of an entity',
-            # TODO make proper JSON-LD definition
-            'path': 'path name of an entity relative to the searched base dataset',
-            # TODO make proper JSON-LD definition
-            'parentds': 'path of the datasets that contains an entity',
-            # 'type' will not come from a metadata field, hence will not be detected
-            'type': 'type of a record',
-        }
-
-        schema_fields = {
-            n.lstrip('@'): wf.ID(stored=True, unique=n == '@id')
-            for n in definitions}
-
-        lgr.debug('Scanning for metadata keys')
-        # quick 1st pass over all dataset to gather the needed schema fields
-        log_progress(
-            lgr.info,
-            'idxschemabuild',
-            'Start building search schema',
-            total=len(dsinfo),
-            label='Building search schema',
-            unit=' Datasets',
-        )
-        for res in query_aggregated_metadata(
-                # XXX TODO After #2156 datasets may not necessarily carry all
-                # keys in the "unique" summary
-                reporton='datasets',
-                ds=self.ds,
-                aps=[dict(path=self.ds.path, type='dataset')],
-                recursive=True):
-            meta = res.get('metadata', {})
-            # no stringification of values for speed, we do not need/use the
-            # actual values at this point, only the keys
-            idxd = _meta2autofield_dict(meta, val2str=False)
-
-            for k in idxd:
-                schema_fields[k] = wf.TEXT(stored=False,
-                                           analyzer=SimpleAnalyzer())
-            log_progress(lgr.info, 'idxschemabuild',
-                         'Scanned dataset at %s', res['path'],
-                         update=1, increment=True)
-        log_progress(
-            lgr.info, 'idxschemabuild', 'Done building search schema')
-
-        self.schema = wf.Schema(**schema_fields)
-
-    def _mk_parser(self):
-        from whoosh import qparser as qparse
-
-        parser = qparse.MultifieldParser(
-            self.idx_obj.schema.names(),
-            self.idx_obj.schema)
-        # XXX: plugin is broken in Debian's whoosh 2.7.0-2, but already fixed
-        # upstream
-        parser.add_plugin(qparse.FuzzyTermPlugin())
-        parser.add_plugin(qparse.GtLtPlugin())
-        parser.add_plugin(qparse.SingleQuotePlugin())
-        # replace field defintion to allow for colons to be part of a field's name:
-        parser.replace_plugin(qparse.FieldsPlugin(expr=r"(?P<text>[()<>.\w]+|[*]):"))
-        self.parser = parser
-
-
-class _EGrepCSSearch(_Search):
-    _mode_label = 'egrepcs'
-    _default_documenttype = 'datasets'
-
-    def __init__(self, ds, **kwargs):
-        super(_EGrepCSSearch, self).__init__(ds, **kwargs)
-        self._queried_keys = None  # to be memoized by get_query
-
-    # If there were custom "per-search engine" options, we could expose
-    # --consider_ucn - search through unique content properties of the dataset
-    #    which might be more computationally demanding
-    def __call__(self, query, max_nresults=None, consider_ucn=False, full_record=True):
-        if max_nresults is None:
-            # no limit by default
-            max_nresults = 0
-        query = self.get_query(query)
-
-        nhits = 0
-        for res in query_aggregated_metadata(
-                reporton=self.documenttype,
-                ds=self.ds,
-                aps=[dict(path=self.ds.path, type='dataset')],
-                # MIH: I cannot see a case when we would not want recursion (within
-                # the metadata)
-                recursive=True):
-            # this assumes that files are reported after each dataset report,
-            # and after a subsequent dataset report no files for the previous
-            # dataset will be reported again
-            meta = res.get('metadata', {})
-            # produce a flattened metadata dict to search through
-            doc = _meta2autofield_dict(meta, val2str=True, consider_ucn=consider_ucn)
-            # inject a few basic properties into the dict
-            # analog to what the other modes do in their index
-            doc.update({
-                k: res[k] for k in ('@id', 'type', 'path', 'parentds')
-                if k in res})
-            # use search instead of match to not just get hits at the start of the string
-            # this will be slower, but avoids having to use actual regex syntax at the user
-            # side even for simple queries
-            # DOTALL is needed to handle multiline description fields and such, and still
-            # be able to match content coming for a later field
-            lgr.log(7, "Querying %s among %d items", query, len(doc))
-            t0 = time()
-            matches = {(q['query'] if isinstance(q, dict) else q, k):
-                       q['query'].search(v) if isinstance(q, dict) else q.search(v)
-                       for k, v in doc.items()
-                       for q in query
-                       if not isinstance(q, dict) or q['field'].match(k)}
-            dt = time() - t0
-            lgr.log(7, "Finished querying in %f sec", dt)
-            # retain what actually matched
-            matched = {k[1]: match.group() for k, match in matches.items() if match}
-            # implement AND behavior across query expressions, but OR behavior
-            # across queries matching multiple fields for a single query expression
-            # for multiple queries, this makes it consistent with a query that
-            # has no field specification
-            if matched and len(query) == len(set(k[0] for k in matches if matches[k])):
-                hit = dict(
-                    res,
-                    action='search',
-                    query_matched=matched,
-                )
-                yield hit
-                nhits += 1
-                if max_nresults and nhits == max_nresults:
-                    # report query stats
-                    topstr = '{} top {}'.format(
-                        max_nresults,
-                        single_or_plural('match', 'matches', max_nresults)
-                    )
-                    lgr.info(
-                        "Reached the limit of {}, there could be more which "
-                        "were not reported.".format(topstr)
-                    )
-                    break
-
-    def show_keys(self, mode=None, regexes=None):
-        """
-
-        Parameters
-        ----------
-        mode: {"name", "short", "full"}
-        regexes: list of regex
-          Which keys to bother working on
-        """
-        maxl = 100  # approx maximal line length for unique values in mode=short
-        # use a dict already, later we need to map to a definition
-        # meanwhile map to the values
-
-        keys = self._get_keys(mode)
-
-        for k in sorted(keys):
-            if regexes and not self._key_matches(k, regexes):
-                continue
-            if mode == 'name':
-                print(k)
-                continue
-
-            # do a bit more
-            stat = keys[k]
-            all_uvals = uvals = sorted(stat.uvals)
-
-            stat.uvals_str = assure_unicode(
-                "{} unique values: ".format(len(all_uvals))
-            )
-
-            if mode == 'short':
-                # show only up until we fill maxl
-                uvals_str = ''
-                uvals = []
-                for v in all_uvals:
-                    appendix = ('; ' if uvals else '') + v
-                    if len(uvals_str) + len(appendix) > maxl - len(stat.uvals_str):
-                        break
-                    uvals.append(v)
-                    uvals_str += appendix
-            elif mode == 'full':
-                pass
-            else:
-                raise ValueError(
-                    "Unknown value for stats. Know full and short")
-
-            stat.uvals_str += '; '.join(uvals)
-
-            if len(all_uvals) > len(uvals):
-                stat.uvals_str += \
-                    '; +%s' % single_or_plural("value", "values", len(all_uvals) - len(uvals), True)
-
-            print(
-                u'{k}\n in  {stat.ndatasets} datasets\n has {stat.uvals_str}'.format(
-                k=k, stat=stat
-            ))
-        # After #2156 datasets may not necessarily carry all
-        # keys in the "unique" summary
-        lgr.warning('In this search mode, the reported list of metadata keys may be incomplete')
-
-
-    def _get_keys(self, mode=None):
-        """Return keys and their statistics if mode != 'name'."""
-        class key_stat:
-            def __init__(self):
-                self.ndatasets = 0  # how many datasets have this field
-                self.uvals = set()
-
-        from collections import defaultdict
-        keys = defaultdict(key_stat)
-        for res in query_aggregated_metadata(
-                # XXX TODO After #2156 datasets may not necessarily carry all
-                # keys in the "unique" summary
-                reporton='datasets',
-                ds=self.ds,
-                aps=[dict(path=self.ds.path, type='dataset')],
-                recursive=True):
-            meta = res.get('metadata', {})
-            # inject a few basic properties into the dict
-            # analog to what the other modes do in their index
-            meta.update({
-                k: res.get(k, None) for k in ('@id', 'type', 'path', 'parentds')
-                # parentds is tricky all files will have it, but the dataset
-                # queried above might not (single dataset), let's force it in
-                if k == 'parentds' or k in res})
-
-            # no stringification of values for speed
-            idxd = _meta2autofield_dict(meta, val2str=False)
-
-            for k, kvals in idxd.items():
-                # TODO deal with conflicting definitions when available
-                keys[k].ndatasets += 1
-                if mode == 'name':
-                    continue
-                keys[k].uvals |= self.get_repr_uvalues(kvals)
-        return keys
-
-    def get_repr_uvalues(self, kvals):
-        kvals_set = set()
-        if not kvals:
-            return kvals_set
-        kvals_iter = (
-            kvals
-            if hasattr(kvals, '__iter__') and not isinstance(kvals, (str, bytes))
-            else [kvals]
-        )
-        return set(shortened_repr(x, 50) for x in kvals_iter)
-
-    def get_query(self, query):
-        query = assure_list(query)
-        simple_fieldspec = re.compile(r"(?P<field>\S*?):(?P<query>.*)")
-        quoted_fieldspec = re.compile(r"'(?P<field>[^']+?)':(?P<query>.*)")
-        query_rec_matches = [
-            simple_fieldspec.match(q) or
-            quoted_fieldspec.match(q) or
-            q
-            for q in query]
-        query_group_dicts_only = [
-            q.groupdict() for q in query_rec_matches if hasattr(q, 'groupdict')
-        ]
-        self._queried_keys = [
-            qgd['field']
-            for qgd in query_group_dicts_only
-            if ('field' in qgd and qgd['field'])
-        ]
-        if len(query_group_dicts_only) != len(query_rec_matches):
-            # we had a query element without field specification add
-            # None as an indicator of that
-            self._queried_keys.append(None)
-        # expand matches, compile expressions
-        query = [
-            {k: self._compile_query(v) for k, v in q.groupdict().items()}
-            if hasattr(q, 'groupdict') else self._compile_query(q)
-            for q in query_rec_matches
-        ]
-
-        # turn "empty" field specs into simple queries
-        # this is used to forcibly disable field-based search
-        # e.g. when searching for a value
-        query = [q['query']
-                 if isinstance(q, dict) and q['field'].pattern == '' else q
-                 for q in query]
-        return query
-
-    def _xfm_query(self, q):
-        # implement potential transformations of regex before they get compiled
-        return q
-
-    def _compile_query(self, q):
-        """xfm and compile the query, with informative exception if query is incorrect
-        """
-        q_xfmed = self._xfm_query(q)
-        try:
-            return re.compile(q_xfmed)
-        except re.error as exc:
-            omsg = " (original: '%s')" % q if q != q_xfmed else ''
-            raise ValueError(
-                "regular expression '%s'%s is incorrect: %s"
-                % (q_xfmed, omsg, exc)
-            )
-
-    def get_nohits_msg(self):
-        """Given the query and performed search, provide recommendation
-
-        Quite often a key in the query is mistyped or I simply query for something
-        which is not actually known.  It requires --show-keys  run first, doing
-        visual search etc to mitigate.  Here we can analyze either all queried
-        keys are actually known, and if not known -- what would be the ones available.
-
-        Returns
-        -------
-        str
-          A sentence or a paragraph to be logged/output
-        """
-        #
-        queried_keys = self._queried_keys[:]
-        if queried_keys and None in queried_keys:
-            queried_keys.pop(queried_keys.index(None))
-        if not queried_keys:
-            return  # No keys were queried, we are of no use here
-        known_keys = self._get_keys(mode='name')
-        unknown_keys = sorted(list(set(queried_keys).difference(known_keys)))
-        if not unknown_keys:
-            return  # again we are of no help here
-        msg = super(_EGrepCSSearch, self).get_nohits_msg()
-        msg += " Following keys were not found in available metadata: %s. " \
-              % ", ".join(unknown_keys)
-        suggestions_msg = get_suggestions_msg(unknown_keys, known_keys)
-        if suggestions_msg:
-            msg += ' ' + suggestions_msg
-        return msg
-
-
-class _EGrepSearch(_EGrepCSSearch):
-    _mode_label = 'egrep'
-    _default_documenttype = 'datasets'
-
-    def _xfm_query(self, q):
-        if q == q.lower():
-            # we have no upper case symbol in the query, go case-insensitive
-            return '(?i){}'.format(q)
-        else:
-            return q
-
-
-class _PyEvalSearch(_Search):
-    """A search engine which consumes a query
-    which is a Python code to be evaluated in an environment where metadata
-    keys are provided as variables (@id -> id).
-    """
-    _mode_label = 'pyeval'
-    _default_documenttype = 'datasets'
-
-    # def __init__(self, ds, **kwargs):
-    #     super(_PyEvalSearch, self).__init__(ds, **kwargs)
-    #     #self._queried_keys = None  # to be memoized by get_query
-
-    # TODO: I feel that we need some base class/mixin like a
-    #  _LoopySearch which would implement this main loop in a single location
-    #  with uniform treatment of metadata (adding those fields into doc etc)
-    # TODO2: that loop whould be two nested dataset/files so we could
-    #  quickly exit whenever enough per dataset (e.g. 1) found
-    def __call__(self, query, max_nresults=None, full_record=True):
-        query = assure_list(query)
-        if max_nresults is None:
-            # no limit by default
-            max_nresults = 0
-        # separate ones are considered to be independent clauses for "and"
-        composite_query = ' and '.join(['(%s)' % q for q in query])
-        composite_query_ = compile(composite_query, 'query', 'eval')
-
-        def matcher(doc):
-            try:
-                return eval(composite_query_, {}, doc)
-            except (KeyError, NameError) as exc:
-                # lgr.debug("Something was missing in the record -- should be generally ok")
-                # TODO: record if ANY matching resulted in not hitting this.
-                #  If not -- we must have misspecified something, and should
-                #  alert user similarly to get_nohits_msg of egrep.
-                #  So -- we need smth like _queried_keys here as well
-                pass
-            except Exception as exc:
-                # TODO: we could analyze each part of composite_query independently
-                # (if len(query) > 1) to provide a more specific pointer to what part
-                # has caused a failure
-                lgr.info("Failed to match %s: %s", query, exc_str())
-
-        nhits = 0
-        for res in query_aggregated_metadata(
-                reporton=self.documenttype,
-                ds=self.ds,
-                aps=[dict(path=self.ds.path, type='dataset')],
-                # MIH: I cannot see a case when we would not want recursion (within
-                # the metadata)
-                recursive=True):
-            # this assumes that files are reported after each dataset report,
-            # and after a subsequent dataset report no files for the previous
-            # dataset will be reported again
-            meta = res.get('metadata', {})
-            # produce a flattened metadata dict to search through
-            doc = meta.copy() #RF _meta2autofield_dict(meta, val2str=True, consider_ucn=consider_ucn)
-            # inject a few basic properties into the dict
-            # analog to what the other modes do in their index
-            doc.update({
-                # RF: @ replacement specific to here
-                k.replace('@', ''): res[k] for k in ('@id', 'type', 'path', 'parentds')
-                if k in res})
-            # use search instead of match to not just get hits at the start of the string
-            # this will be slower, but avoids having to use actual regex syntax at the user
-            # side even for simple queries
-            # DOTALL is needed to handle multiline description fields and such, and still
-            # be able to match content coming for a later field
-            lgr.log(7, "Querying %s among %d items", query, len(doc))
-            t0 = time()
-            is_a_match = matcher(doc)
-            dt = time() - t0
-            lgr.log(7, "Finished querying in %f sec", dt)
-            # The rest below seemed to be specific to the grep one, although
-            # TODO: check exactly what was intended
-            # # retain what actually matched
-            # matched = {k[1]: match.group() for k, match in matches.items() if match}
-            # # implement AND behavior across query expressions, but OR behavior
-            # # across queries matching multiple fields for a single query expression
-            # # for multiple queries, this makes it consistent with a query that
-            # # has no field specification
-            # if matched and len(query) == len(set(k[0] for k in matches if matches[k])):
-            if is_a_match:
-                hit = dict(
-                    res,
-                    action='search',
-                    # query_matched=matched,
-                )
-                yield hit
-                nhits += 1
-                if max_nresults and nhits == max_nresults:
-                    # report query stats
-                    topstr = '{} top {}'.format(
-                        max_nresults,
-                        single_or_plural('match', 'matches', max_nresults)
-                    )
-                    lgr.info(
-                        "Reached the limit of {}, there could be more which "
-                        "were not reported.".format(topstr)
-                    )
-                    break
-
 
 
 @build_doc
